refactor(tools): extract value-set constants to leaf module
Prep for the E1 tool registry: TRANSACTION_TYPE_FILTERS, CATEGORY_VIEWS,
and BALANCE_HISTORY_GRANULARITIES move to src/tools/constants.ts (a leaf
module with no imports) so upcoming registry modules can share them with
tools.ts without a runtime import cycle. Re-exported from tools.ts for
compatibility — no behavior change.
